api/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1